Changed ssize_t to intmax_t and removed extern keyword for functions.

+++ c/jvm/src/test/scala/ext/ext.c
@@ -2,6 +2,6 @@
 #include <unistd.h>
 
 void top_Console_read(STACK_FRAME IS_541E35 result) {
-  ssize_t n = read(STDIN_FILENO, &result->value, MaxIS_541E35);
+  intmax_t n = read(STDIN_FILENO, &result->value, MaxIS_541E35);
   result->size = (IS_541E35SizeT) n;
 }
